finished program - 1st submission

+++ week04/OnlineOrdering/Product.cs
@@ -2,10 +2,10 @@ public class Product
 {
     private string _name;
     private int _id;
-    private float _price;
+    private double _price;
     private int _quantity;
 
-    public Product(string name, int id, float price, int quantity)
+    public Product(string name, int id, double price, int quantity)
     {
         _name = name;
         _id = id;
